Delete non-necessary test harness interfaces.
