+++ src/main/kotlin/org/wfanet/measurement/loadtest/frontend/BUILD.bazel
@@ -15,7 +15,6 @@ kt_jvm_library(
     name = "frontend_simulator",
     srcs = [
         "FrontendSimulator.kt",
-        "FrontendSimulatorImpl.kt",
     ],
     deps = [
         "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:resource_key",
